added recentfiles.cpp to vs2005 project, fixed some warnings, removed Release()-call that caused crash.

+++ editor/trackview.h
@@ -27,7 +27,7 @@ public:
 	SyncDocument *getDocument() { return document; }
 	
 	void setRows(size_t rows);
-	int getRows() const
+	size_t getRows() const
 	{
 		if (NULL == document) return 0;
 		return document->getRows();
